Registers geolocator plugin for macOS
Registers the geolocator plugin with the Flutter plugin registry for macOS.

This allows the Flutter application to use the geolocator plugin to access location services on macOS.

Registers geolocator plugin

Registers the geolocator plugin with the Flutter plugin registry
for macOS and Windows.

This allows Flutter applications to use the geolocator plugin
to access location services on these platforms.